requirements.txt
Django>=1.11.6,<2.0
django-compressor==2.0
django-crispy-forms==1.6.0
django-extensions==1.6.1
django-filter==1.1.0
django-guardian==1.4.9
django-js-utils==0.0.5.dev0
django-sendfile==0.3.10
djangorestframework==3.7.1
djangorestframework-csv==2.0.0
djangorestframework-filters==0.10.2
jsonfield==1.0.3
psycopg2
rcssmin==1.0.6
rjsmin==1.0.12
tablib==0.11.2
